+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # PythonQwt Releases
 
+## Version 0.11.1 (work in progress)
+
+- ...
+
 ## Version 0.11.0
 
 - Dropped support for Python 3.7 and earlier